Changed Color_RGBA8_u32 struct (#163)
* Changed `Color_RGBA8_u32` struct.

* Removed one color class and a replaced it with the correct one.

* Fixed the color issue in GfxPrint

+++ include/def/gfxprint.h
@@ -7,6 +7,6 @@ void GfxPrint_Init(GfxPrint* pthis);
 void GfxPrint_Open(GfxPrint* pthis, Gfx* dList);
 s32 GfxPrint_Printf(GfxPrint* pthis, const char* fmt, ...);
 void GfxPrint_SetBasePosPx(GfxPrint* pthis, s32 x, s32 y);
-void GfxPrint_SetColor(GfxPrint* pthis, u32 r, u32 g, u32 b, u32 a);
+void GfxPrint_SetColor(GfxPrint* pthis, u8 r, u8 g, u8 b, u8 a);
 void GfxPrint_SetPos(GfxPrint* pthis, s32 x, s32 y);
 void GfxPrint_SetPosPx(GfxPrint* pthis, s32 x, s32 y);
